requirements.txt
numpy==1.24.4
pandas==2.0.3

geopandas==0.13.2
shapely==2.0.1
pyproj==3.5.0
geopy==2.4.1

matplotlib==3.7.5
scikit-learn==1.3.2
optuna==3.4.0

python-dateutil==2.9.0.post0
requests==2.31.0

clearml